feat: file and folder stuct
